AsyncTask: Use a mutex to guard access to AsyncTask.status_line.
While a backup or restore task was running, the status_line property
would be updated for each file processed.

Meanwhile, the gui (BackupBox, for example) was periodically
accessing the status_line property to set a label.

Without protection, the gui was picking up garbage from status_line
instead of the filename, and was flooding output with markup warnings.
There was probably potential for a crash as well.