Fix project's license
The source code is licensed under GPL-2.0-or-later license but
LICENSE.md presents LGPL-3.0-or-later license that is more suitable
for libraries.

So replace LICENSE.md with GPL-2.0-or-later.txt file and place it
under LICENSES directory as suggested by the REUSE standard.

Also remove COPYING as it contains GPL-3.0-or-later license.