Update timeshift.appdata.xml (#356)
The metainfo file put xml:lang on the surrounding <description/> element, not the <p/> elements.

(The specification says: In MetaInfo files, this tag (<description/>) should be translated by-paragraph. For enumerations, items are translated individually as well, and not the whole enumeration block. This means that in a translated file, only <p/> and <li/>elements may carry an xml:lang property.)