Subvolume.vala: (btrfs) Wait for async file operations to complete.
With quotas disabled, available space wasn't being updated
consistently when removing old snapshots.

With quotas enabled, no time was allowed after committing to
delete a subvolume before attempting to remove its corresponding
qgroup, resulting in an error and requiring the user to delete the
snapshot twice.

Use the return codes from subvolume-sync and quota-rescan commands
to control the pace during a delete operation. This allows qgroup-
destroy to succeed and keeps the 'available space' value in the
main window in sync with the current filesystem state.

Unfortunately, as a result of this, snapshot deletion may take
longer to complete even with quotas disabled.

ref: #291