AppStream metadata fixes (#391)
* Remove 'English' marker from AppStream description

appstreamcli seems to require the existence of a description without
a 'xml:lang' attribute. This makes sense if you think about the
"lang-less" description as the default one, and the translated ones
as overrides.

* Reorder description paragraphs

The AppStream metadata spec states that the description "should be
translated by-paragraph". Looking at other metainfo files, it seems
that most treat this as meaning "translations should immediately
follow the non-translated paragraph".

* Add <launchable> tag to AppStream metadata

This tag informs AppStream consumers how to launch the application.

* Change AppStream ID to com.linuxmint.timeshift

AppStream IDs are recommended to follow an reverse-DNS scheme.