improve retrieving of session related environment variables (#460)
instead of using env in the timeshift-launcher like proposed in #384, the timeshift process itself now searches its parents until it finds a process that is not owned by the same user as timeshift (root) and then copies interesting environment variables from there. This allows pkexec to handle the command correctly.